fix: cdek integration

+++ app/modules/delivery/schemas/create_order.py
@@ -7,16 +7,16 @@ class CdekPackageItem(BaseModel):
     ware_key: str
     name: str
     cost: float
-    weight: int
-    amount: int
+    weight: float
+    amount: float
     payment: dict
 
 class CdekPackage(BaseModel):
     number: str
-    weight: int
-    length: int
-    width: int
-    height: int
+    weight: float
+    length: float
+    width: float
+    height: float
     items: tp.List[CdekPackageItem]
 
 class CdekRecipient(BaseModel):
